scripts: west_commands: refactor run_common.py
Pull out some more common functionality we will need elsewhere into a
separate file.
